WoT restructure: dissolve cellLocator/ -> cellAccess to table/ (table-only), cellCoords to utils/ (event-only)
