+++ handsontable/src/3rdparty/walkontable/src/table/baseTable.ts
@@ -2,7 +2,7 @@ import type { WalkontableInstance } from '../types';
 import type { EngineContext } from '../wire';
 import type Settings from '../settings';
 import type { RowRangeQuery, ColumnRangeQuery } from '../rangeQuery/renderedRange';
-import { cellAccess, type CellAccess } from '../cellLocator/cellAccess';
+import { cellAccess, type CellAccess } from './cellAccess';
 import { domScaffold, type DomScaffold } from './domScaffold';
 import {
   isHTMLElement,
