+++ handsontable/src/3rdparty/walkontable/src/utils/cellCoords.ts
